pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-mail>=0.10.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "wtforms>=3.2.1",
    "pyjwt>=2.10.1",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
    "werkzeug>=3.1.3",
    "pandas>=2.3.2",
    "openpyxl>=3.1.5",
    "sqlalchemy>=2.0.42",
    "openai>=1.107.0",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
    "numpy>=2.3.2",
    "rembg>=2.0.67",
    "onnxruntime>=1.23.2",
    "rapidfuzz>=3.14.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "qrcode>=8.2",
    "scipy>=1.16.3",
    "gevent>=25.9.1",
    "flask-swagger-ui>=5.32.1.post2",
    "bleach>=6.3.0",
    "markupsafe>=3.0.2",
    "pywebpush>=2.3.0",
]
